+++ CHANGELOG.md
@@ -2,6 +2,10 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ## [Unreleased]
 
+### Enhancements
+
+* Faster `nrt` using Newton's method
+
 ### v25.9.13
 
 ### Enhancements